headset: fix AT +XAPL messages reponse for devices that use APPLE protocol
This fixes the battery reporting on devices using the AT command +XAPL.
The stock implementation reponds with a hardcoded "2" flag, asking only
for battery events. The device expects a response with the same flag sent
in the initial negotiation message, so let's use that instead of the
hardcoded "2" and be done with it. Tested on a Plantronics PLT_E50 and
the battery reporting is fully working now.

Change-Id: I2f17dec7a96b70132ce21260902b74e89bd7f332